+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+
 ## 0.0.8
 - Remove semantic highlighting as Quarto extension already provide it
 - Add a `Insert Code Chunk` command that allows user to insert a code chunk with a language of their choice.