+++ CHANGELOG.md
@@ -1,27 +1,38 @@
 # Change Log
 
+## 0.0.9
+
+- Fix issue related to Zotero citation does not work
 
 ## 0.0.8
+
 - Remove semantic highlighting as Quarto extension already provide it
 - Add a `Insert Code Chunk` command that allows user to insert a code chunk with a language of their choice.
 
 ## 0.0.7
+
 - Add a button to cite document from a runtime Zotero
 
 ## 0.0.6
+
 - Cite reference directly from a Zotero runtime and update the `*.bib` file. Default is to the `*.bib` file relative to the workspace root. Command `Quarto Wingman: Select citation from Zotero`
 
 ## 0.0.5
+
 - Fix problem when running inline code for VSCode user
 
 ## 0.0.4
+
 - Update new logo
 
 ## 0.0.3
+
 - Disable `Run Inline Code` for VSCode for future investigation
 
 ## 0.0.2
+
 - Fix the error the codelens not displaying
 
 ## 0.0.1
+
 - Initial release